move utility methods to their own directory
