+++ ten_k_brews_app/utilities/detect_mobile.py
